refactor: group all badges under a Components/badges folder
