+++ web/src/components/badges/SeverityBadge/SeverityBadge.tsx
