+++ web/src/components/badges/SeverityBadge/index.tsx
