+++ web/src/pages/AlertDetails/AlertDetailsInfo/AlertDetailsInfo.tsx
@@ -23,7 +23,7 @@ import { RuleDetails } from 'Generated/schema';
 import Linkify from 'Components/Linkify';
 import { formatDatetime } from 'Helpers/utils';
 import { Link as RRLink } from 'react-router-dom';
-import SeverityBadge from 'Components/SeverityBadge';
+import SeverityBadge from 'Components/badges/SeverityBadge';
 import UpdateAlertDropdown from 'Components/dropdowns/UpdateAlertDropdown';
 import { AlertSummaryFull } from 'Source/graphql/fragments/AlertSummaryFull.generated';
 import { AlertDetailsFull } from 'Source/graphql/fragments/AlertDetailsFull.generated';
