+++ web/src/pages/LogAnalysisOverview/AlertsTable/AlertsTable.tsx
@@ -20,10 +20,10 @@ import React from 'react';
 import { Box, Link, Table, Icon } from 'pouncejs';
 import urls from 'Source/urls';
 import { Link as RRLink } from 'react-router-dom';
-import SeverityBadge from 'Components/SeverityBadge';
+import SeverityBadge from 'Components/badges/SeverityBadge';
 import { ListAlerts } from 'Pages/ListAlerts/graphql/listAlerts.generated';
 import { shortenId, formatDatetime } from 'Helpers/utils';
-import AlertStatusBadge from 'Components/AlertStatusBadge';
+import AlertStatusBadge from 'Components/badges/AlertStatusBadge';
 
 type ListAlertsTableProps = {
   items: ListAlerts['alerts']['alertSummaries'];
