+++ web/src/pages/RuleDetails/RuleDetailsAlertsTable/RuleDetailsAlertsTable.tsx
@@ -21,7 +21,7 @@ import { Box, Link, Table } from 'pouncejs';
 import { Link as RRLink } from 'react-router-dom';
 import urls from 'Source/urls';
 import { formatDatetime, shortenId } from 'Helpers/utils';
-import SeverityBadge from 'Components/SeverityBadge';
+import SeverityBadge from 'Components/badges/SeverityBadge';
 import UpdateAlertDropdown from 'Components/dropdowns/UpdateAlertDropdown';
 import { ListAlertsForRule } from '../graphql/listAlertsForRule.generated';
